+++ operator/pkg/resources/cmn/certificate_internal_test.go
@@ -19,7 +19,7 @@ var _ = Describe("buildCertificateSANs", func() {
 				Namespace: "test-ns",
 			},
 			Spec: aisv1.AIStoreSpec{
-				TLSCertificate: &aisv1.TLSCertificateSpec{
+				TLSCertificate: &aisv1.TLSCertificateConfig{
 					AdditionalDNSNames: []string{"test-additional-dns-name"},
 				},
 				HostnameMap: map[string]string{